+++ src/button-accordion-with-keyboard/atoms/buttonStates.ts
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const buttonStatesAtom = atom<Record<string, boolean>>({});
